Extract form data to components/Forms
