+++ assets/js/components/Forms/FormContext.tsx
@@ -2,9 +2,9 @@ import * as React from "react";
 
 import type { FormState } from "./FormState";
 
-export const FormContext = React.createContext<FormState | null>(null);
+export const FormContext = React.createContext<FormState<any> | null>(null);
 
-export function getFormContext(): FormState {
+export function getFormContext(): FormState<any> {
   const form = React.useContext(FormContext);
   if (!form) throw new Error("Form fields must be used within a Form component");
   return form;
