+++ assets/js/components/Pages/index.tsx
@@ -2,3 +2,5 @@ export { Page } from "./Page";
 
 export { useRefresh } from "./useRefresh";
 export { useLoadedData } from "./useLoadedData";
+
+export const emptyLoader = () => Promise.resolve(null);
